# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/memberships": {
      "get": {
        "operationId": "MembershipsController_findAll",
        "summary": "List all memberships, optionally filtered by communityId or userId",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role for RBAC: admin | community_manager | moderator | user",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "communityId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": [
                    {
                      "id": 1,
                      "userId": 3,
                      "communityId": 1,
                      "joinedAt": "2026-01-15T11:30:00.000Z"
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "memberships"
        ]
      },
      "post": {
        "operationId": "MembershipsController_create",
        "summary": "Join a community (create membership)",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role for RBAC: admin | community_manager | moderator | user",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMembershipDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": 5,
                    "userId": 3,
                    "communityId": 1,
                    "joinedAt": "2026-05-02T12:00:00.000Z"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "memberships"
        ]
      }
    },
    "/api/memberships/{id}": {
      "delete": {
        "operationId": "MembershipsController_remove",
        "summary": "Leave a community (delete membership by id)",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role for RBAC: admin | community_manager | moderator | user",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Membership 5 removed"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "memberships"
        ]
      }
    },
...
    "/api/channels/{id}/messages": {
      "get": {
        "operationId": "MessagesController_findByChannel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "messages"
        ]
      }
    },
    "/api/messages": {
      "post": {
        "operationId": "MessagesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMessageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "messages"
        ]
      }
    },
    "/api/messages/{id}/reactions": {
      "post": {
        "operationId": "MessagesController_react",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReactMessageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "messages"
        ]
      }
    },
    "/api/messages/{id}/pin": {
      "patch": {
        "operationId": "MessagesController_pin",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "messages"
        ]
      }
    },
    "/api/messages/{id}/attachments": {
      "post": {
        "operationId": "MessagesController_addAttachment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddMessageAttachmentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "messages"
        ]
      }
    },
...
    "/api/notifications": {
      "get": {
        "operationId": "NotificationsController_findAll",
        "summary": "List notifications for a user, newest first",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Array of notifications for the given userId"
          }
        },
        "tags": [
          "notifications"
        ]
      }
    },
    "/api/notifications/read-all": {
      "patch": {
        "operationId": "NotificationsController_markAllRead",
        "summary": "Mark every notification for a user as read",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarkAllReadDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "notifications"
        ]
      }
    },
    "/api/notifications/{id}": {
      "patch": {
        "operationId": "NotificationsController_update",
        "summary": "Mark a single notification read/unread",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNotificationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "notifications"
        ]
      }
    },
...
    "/api/moderator-certification": {
      "get": {
        "operationId": "ModeratorCertificationController_findAll",
        "summary": "List all moderator applications",
        "description": "Admin-only — powers the approval queue.",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "RBAC role header. quiz/apply/status: user | organizer | admin. PATCH :id (certify/reject) requires admin only.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "pending",
                "certified",
                "rejected"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "moderator-certification"
        ]
      }
    },
    "/api/moderator-certification/quiz": {
      "get": {
        "operationId": "ModeratorCertificationController_quiz",
        "summary": "Get the moderation quiz questions",
        "description": "Answer key is never returned — score via POST /apply.",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "RBAC role header. quiz/apply/status: user | organizer | admin. PATCH :id (certify/reject) requires admin only.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "moderator-certification"
        ]
      }
    },
    "/api/moderator-certification/apply": {
      "post": {
        "operationId": "ModeratorCertificationController_apply",
        "summary": "Submit the moderation quiz to apply for Certified Moderator status",
        "description": "Scoring ≥70% moves to pending admin review; below 70% is rejected immediately and may be retaken.",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "RBAC role header. quiz/apply/status: user | organizer | admin. PATCH :id (certify/reject) requires admin only.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitQuizDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "moderator-certification"
        ]
      }
    },
    "/api/moderator-certification/{id}": {
      "patch": {
        "operationId": "ModeratorCertificationController_setStatus",
        "summary": "Certify or reject a moderator application",
        "description": "Admin-only. On certification, also flips the applicant's account role to moderator.",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "RBAC role header. quiz/apply/status: user | organizer | admin. PATCH :id (certify/reject) requires admin only.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateModeratorStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "moderator-certification"
        ]
      }
    },
    "/api/moderator-certification/status": {
      "get": {
        "operationId": "ModeratorCertificationController_status",
        "summary": "Get a user's moderator application status",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "RBAC role header. quiz/apply/status: user | organizer | admin. PATCH :id (certify/reject) requires admin only.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "schema": {
              "example": 4,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Application status and quiz score"
          }
        },
        "tags": [
          "moderator-certification"
        ]
      }
    },
...
  },
  "info": {
    "title": "Gameunity Backend API",
    "description": "In-memory NestJS backend for frontend integration and academic evaluation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Dev Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "x-role": {
        "type": "apiKey",
        "in": "header",
        "name": "x-role",
        "description": "RBAC role header: admin | community_manager | moderator | user"
      }
    },
    "schemas": {
...
      "CreateMembershipDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "number",
            "example": 3,
            "description": "ID of the user joining the community"
          },
          "communityId": {
            "type": "number",
            "example": 1,
            "description": "ID of the community to join"
          }
        },
        "required": [
          "userId",
          "communityId"
        ]
      },
...
      "CreateMessageDto": {
        "type": "object",
        "properties": {
          "channelId": {
            "type": "string",
            "example": "general"
          },
          "communityId": {
            "type": "number",
            "example": 1
          },
          "authorId": {
            "type": "number",
            "example": 4
          },
          "authorName": {
            "type": "string",
            "example": "player01"
          },
          "content": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2000,
            "example": "Anyone queueing ranked tonight?"
          },
          "attachments": {
            "example": [],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "channelId",
          "content"
        ]
      },
      "ReactMessageDto": {
        "type": "object",
        "properties": {
          "emoji": {
            "type": "string",
            "example": "👍"
          },
          "actorId": {
            "type": "number",
            "description": "Who is reacting — used only to notify the message author",
            "example": 5
          },
          "actorName": {
            "type": "string",
            "example": "player01"
          }
        }
      },
      "AddMessageAttachmentDto": {
        "type": "object",
        "properties": {
          "fileName": {
            "type": "string",
            "example": "screenshot.png"
          },
          "url": {
            "type": "string",
            "example": "/uploads/1234-screenshot.png"
          }
        }
      },
...
      "MarkAllReadDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "number",
            "example": 5
          }
        },
        "required": [
          "userId"
        ]
      },
      "UpdateNotificationDto": {
        "type": "object",
        "properties": {
          "read": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "read"
        ]
      },
...
      "SubmitQuizDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "number",
            "example": 4
          },
          "answers": {
            "example": [
              1,
              0,
              2,
              1,
              0
            ],
            "description": "Zero-indexed selected option per question, in the order returned by GET /moderator-certification/quiz",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "userId",
          "answers"
        ]
      },
      "UpdateModeratorStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "certified",
            "enum": [
              "certified",
              "rejected"
            ]
          }
        },
        "required": [
          "status"
        ]
      },
...
    }
  }
}